+++ CheckoutLogic/ICheckout.cs
@@ -3,5 +3,5 @@
 internal interface ICheckout
 {
     public void Scan(string item);
-    public int GetTotalPrice();
+    public decimal GetTotalPrice();
 }
